Fixed NotificationManager tests
